From Gerri: Patch adding support for backend-dependent record formatting of PROOF monitoring. This is achieved by introducing a new layer, described by the abstract interface TProofMonSender, with the record format defined in the backend implemenattion (currently TProofMonSenderML, for MonaLisa, and TProofMonSenderSQL, for SQL backends). Currently three types of records are sent: 'summary' (derived from what was currently posted), 'dataset', with entries per dataset processed in the query, and 'files', with entries per file processed in the query. In SQL terms, each of this records corresponds to a different table. Sending of any of the three records can be toggled independently.
Record format have been versionned, so that previous versions of the
'summary' record can be still sent chosing the corresponding version.
The latest version of 'summary' has been augmented with information
about the number of missing files and the ROOT version.

The headers of the relevant methods in TProofMonSenderSQL and
TProofMonSenderML contain the details about the records.

The Web doc is up-to-date.

The patch also add to TSQMonitoring::SendParameters the possibility to
do a bulk insert. This feature is used by TProofMonSenderSQL when 
sending dataset or files information.


git-svn-id: http://root.cern.ch/svn/root/trunk@40699 27541ba8-7e3a-0410-8455-c3a389f83636