From Matthew Strait: The class documentation for TTree, TBranch and TLeaf all say "If leaf name has the form var[nelem], where nelem is a digit, then...".  This frightened me for a moment because if taken literally it would mean arrays can't be larger than 9 elements.  A look at the code reveals that it doesn't really mean "digit" but rather "number".  Or, to be precise, "non-negative integer".  *Whew*.
The attached patch changes "digit" to "non-negative integer" in all relevant
cases that I could find.  It also incidentally fixes some spelling and
punctuation and wraps a few gratuitously long lines of code.

Possibly "non-negative" could be "positive", but TLeaf::GetLeafCounter() at
least does not seem to make an exception for the case of zero.


git-svn-id: http://root.cern.ch/svn/root/trunk@40753 27541ba8-7e3a-0410-8455-c3a389f83636