From Matthew Straits:
If one is at the ROOT command line and hits enter with a parenthesis, bracket or similar open, one is presented with a prompt like: 
end with '}', '@':abort > 
While in this mode, CINT writes input into a fixed length buffer (2kB on my system) using strcpy() without checking for overruns. If too much is entered, it happily stomps on the next bit of memory, leading to either an immediate seg fault or memory corruption that leads to a crash later, often when the ROOT session is closed. 
The attached patch makes the relevant function aware of the buffer length and causes the line contination logic to exit with an error rather than corrupting memory.


git-svn-id: http://root.cern.ch/svn/root/trunk@41109 27541ba8-7e3a-0410-8455-c3a389f83636