To handle each input line in cling we put it into separate memory buffer and this buffer gets FileID (not fully valid one) and then it gets processed. There is a problem with the SourceLocations, because the new coming memory buffers have to have inclusion source location. I.e the exact location where they got #included. In the case of ordinary program everything is #included from the main file.
In our case we cannot do that because we don't know how many buffers we will 
end up with and expand the main memory buffer. Even if we knew that wouldn't 
help because our main memory buffer (as main buffer) doesn't have real FileID
and generating a SourceLocation from that doesn't actually work.

If we generate a SourceLocation from a memory buffer that hasn't come from a 
file we will end up with couple of issues. First, we won't be able to print 
proper inclusion stack. Second, those source locations are considered as 
builtins. Last but not least, we confuse the compiler when it does ordering.
Sometimes the compiler needs to know which file came first. In order to do that
it looks up and tries to find the common parent of two source locations and then
compare the offsets. In our case we don't even have common parent.


Chain properly the memory buffers using virtual file (we don't allocate 
anything) of "infinite (1U << 15U)" size and generate valid source location so 
that clang can compare which input became first. 

   +---------------------+
   | Main memory buffer  |
   +---------------------+
   |  Virtual file SLoc  |
   |    address space    |<-----------------+
   |         ...         |<------------+    |
   |         ...         |             |    |
   |         ...         |<----+       |    |
   |         ...         |     |       |    |
   +~~~~~~~~~~~~~~~~~~~~~+     |       |    |
   |     input_line_1    | ....+.......+..--+
   +---------------------+     |       |   
   |     input_line_2    | ....+.....--+
   +---------------------+     |
   |         ...         |     |
   +---------------------+     |
   |     input_line_N    | ..--+  
   +---------------------+


Fixes the issue with the printing of overloads, if they came from different files (included from memory buffers for example). Needed patch in clang.


git-svn-id: http://root.cern.ch/svn/root/trunk@41653 27541ba8-7e3a-0410-8455-c3a389f83636