From Gerri:
In TSeqCollection::Merge adding a TList of objects without a Merge method (e.g. TObjString)
to the output list did *not* work properly, in the sense that all but one set of objects is lost. 
This is because TSeqCollection::Merge (used by TList) is ignoring non mergeable objects instead 
of just adding them to the original list as we do at higher level in PROOF. 
This patch update the behavior of TSeqCollection::Merge to append the non-mergeable objects.


git-svn-id: http://root.cern.ch/svn/root/trunk@41679 27541ba8-7e3a-0410-8455-c3a389f83636