Apply the patch suggested here: https://savannah.cern.ch/bugs/?94807
With option COL TProfile2D histograms are handled differently because, for this type of 2D
histograms, it is possible to know if an empty bin has been filled or not. So even
if all the bins' contents are positive some empty bins might be painted. And vice versa,
if some bins have a negative content some empty bins might be not painted.


git-svn-id: http://root.cern.ch/svn/root/trunk@44301 27541ba8-7e3a-0410-8455-c3a389f83636