Import revision 44750 from the trunk: From Peter, Jack and Elvin:
Indeed one can drop the unnecessary clearing of cache after looping through the collection of baskets. This affects in no way the prefetching. It was left it there for "historical" reasons, when trying to figure out how all the mechanism was working. The FillBuffer code is now pretty entangled and initially it was thought the prefetcher was using this to get the direction of the reads but no, this can be correctly established even without the clearing of the cache.


git-svn-id: http://root.cern.ch/svn/root/branches/v5-34-00-patches@44752 27541ba8-7e3a-0410-8455-c3a389f83636