Move the code generator in the incremental parser.
This is part of more major change in the design:
There is no point of multiple consumers, which are used to transform the decls.
Now when we have the transaction entity is better to implement the consumers as
transaction transformers. One of the biggest advantages that the transformed 
declaration could be put as top level declaration in the transaction. Eg: 
void wrapper() {
  int i;
}

The declaration extractor extracts the declaration on the global scope and it 
becomes top level decl. With the current model there is no way how this will end
up in the transaction.
Another advantage is that we can catch the template instantiation calls from sema
and put the template instances in the relevant transaction as well.


git-svn-id: http://root.cern.ch/svn/root/trunk@45201 27541ba8-7e3a-0410-8455-c3a389f83636